process_icell8.py: move pipeline commands & tasks into library module 'icell8.pipeline'.
